Introduce HTTPS transport
UDP and TCP are great if the server is within the same secured
network.
TCP with TLS is working if one wants to deal with certificates
at the Graylog server.
HTTPS should be introduced as an option to send logs via the
internet in a encrypted way without necessarily setting up
TCP-TLS.